+++ build.gradle
@@ -11,7 +11,7 @@ buildscript {
     jcenter()
   }
 
-  dependencies {       
+  dependencies {
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
     classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
   }
